+++ lib/src/component/menu_item.dart
@@ -4,6 +4,7 @@ class _MenuItem {
   List<String> list;
   int index;
   String menuTitle;
+  dynamic curMenuData;
 
-  _MenuItem(this.list, this.index, this.menuTitle);
+  _MenuItem(this.list, this.index, this.menuTitle, this.curMenuData);
 }